feat: reworked examples
