+++ frontend/src/examples/.eslintrc.json
@@ -1,0 +1,6 @@
+{
+    "rules": {
+        "@typescript-eslint/no-var-requires": "off",
+        "@typescript-eslint/explicit-function-return-type": "off"
+    }
+}
