+++ frontend/src/examples/evaluate-javascript.js
@@ -1,0 +1,18 @@
+const playwright = require("playwright");
+
+(async () => {
+  const browser = await playwright.webkit.launch();
+  const context = await browser.newContext();
+  const page = await context.newPage();
+  await page.goto('https://www.example.com/');
+  const dimensions = await page.evaluate(() => {
+    return {
+      width: document.documentElement.clientWidth,
+      height: document.documentElement.clientHeight,
+      deviceScaleFactor: window.devicePixelRatio
+    }
+  })
+  console.log(JSON.stringify(dimensions));
+
+  await browser.close();
+})();
