+++ frontend/src/examples/generate-pdf.js
@@ -1,0 +1,10 @@
+const playwright = require("playwright");
+
+(async () => {
+  const browser = await playwright.chromium.launch();
+  const context = await browser.newContext();
+  const page = await context.newPage();
+  await page.goto('https://www.google.com/search?q=Google');
+  await page.pdf({ path: `document.pdf` });
+  await browser.close();
+})();
