+++ frontend/src/examples/mobile-and-geolocation.js
@@ -1,0 +1,19 @@
+const playwright = require("playwright");
+
+(async () => {
+  const { webkit, devices } = playwright;
+  const iPhone11 = devices['iPhone 11 Pro'];
+  const browser = await webkit.launch();
+  const context = await browser.newContext({
+    viewport: iPhone11.viewport,
+    userAgent: iPhone11.userAgent,
+    geolocation: { longitude: 12.492507, latitude: 41.889938 },
+    permissions: ['geolocation']
+  });
+  const page = await context.newPage();
+  await page.goto('https://maps.google.com');
+  await page.click(".ml-my-location-fab button");
+  await page.waitForRequest(/.*preview\/pwa/);
+  await page.screenshot({ path: 'colosseum-iphone.png' });
+  await browser.close();
+})();
