+++ frontend/src/examples/page-screenshot.js
@@ -1,0 +1,13 @@
+const playwright = require("playwright");
+
+(async () => {
+  // Try to add 'firefox' to the list ↓
+  for (const browserType of ['chromium', 'webkit']) {
+    const browser = await playwright[browserType].launch();
+    const context = await browser.newContext();
+    const page = await context.newPage();
+    await page.goto('http://whatsmyuseragent.org/');
+    await page.screenshot({ path: `example-${browserType}.png` });
+    await browser.close();
+  }
+})();
